Refactioring conversation
